Add isVerify parameter and revise for the "cannot delete" course rooms

+++ room/views.py
@@ -9,6 +9,8 @@ from room.serializers import RoomSerializer, RoomMemberSerializer, RoomBlockSeri
     RoomRecordSerializer
 from user.models import CustomUser, Notification
 
+from user.permissions import IsVerify
+
 
 def error_response(message, status_code):
     return Response({"error": message}, status=status_code)
@@ -19,7 +21,7 @@ def room_exist(room_id):
 
 
 class RoomRecordList(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request, room_id):
         records = RoomRecord.objects.filter(room_id=room_id)
@@ -42,7 +44,6 @@ class GetCategoryChoices(APIView):
         response = {key: value for key, value in CATEGORY_CHOICES}
         return JsonResponse(response, safe=False)
 
-
 class RoomList(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
@@ -52,14 +53,13 @@ class RoomList(APIView):
         return Response(serializer.data)
 
     def post(self, request):
+
+        if not request.user.is_verify:
+            return error_response("You do not have the right to create a room", status.HTTP_401_UNAUTHORIZED)
         room_serializer = RoomSerializer(data=request.data)
         data = request.data.copy()
         owner = request.user
-        if data.get("room_type", None) == 'course':
-            data["access_level"] = "manager"
-            owner = CustomUser.objects.get(username="admin")
-        else:
-            data["access_level"] = "admin"
+        data["access_level"] = "admin"
         member_serializer = RoomMemberSerializer(data=data)
 
         if len(RoomMember.objects.filter(member=owner, access_level="admin")) >= 50:
@@ -71,9 +71,7 @@ class RoomList(APIView):
 
         room_serializer.save()
         member_serializer.save(member=request.user, room=room_serializer.instance)
-        if data.get("room_type", None) == 'course':
-            RoomMember(room=room_serializer.instance, member=owner,
-                       nickname="admin", access_level="admin").save()
+
         member = member_serializer.instance
         RoomRecord(room=room_serializer.instance,
                    recording=f"{member.nickname}({member.member.username}) 建立了房間!").save()
@@ -81,7 +79,7 @@ class RoomList(APIView):
 
 
 class RoomDetail(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request, room_id):
         if not room_exist(room_id):
@@ -115,14 +113,17 @@ class RoomDetail(APIView):
             return error_response("You are not in the room", status.HTTP_400_BAD_REQUEST)
         if not RoomMember.objects.get(room_id=room_id, member=request.user).access_level == 'admin':
             return error_response("Only admin is authorized to delete room.", status.HTTP_401_UNAUTHORIZED)
+        if Room.objects.get(id=room_id).room_type == 'course':
+            return error_response("Course rooms cannot be deleted.", status.HTTP_400_BAD_REQUEST)
         # notify users that were in this room
         room = Room.objects.get(id=room_id)
         room.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
+# What's the difference between the 2 following API?
 class RoomMemberList(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request, room_id):
         if not Room.objects.filter(id=room_id).exists():
@@ -133,7 +134,7 @@ class RoomMemberList(APIView):
 
 
 class RoomMemberDetail(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request, room_id, user_id):
         if not room_exist(room_id):
@@ -146,7 +147,7 @@ class RoomMemberDetail(APIView):
 
 
 class RoomJoin(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     # In the case of creating room, it is handled in post() in class RoomList()
     def post(self, request, room_id):
@@ -177,7 +178,7 @@ class RoomJoin(APIView):
 
 
 class RoomLeave(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def delete(self, request, room_id):
         if not Room.objects.filter(id=room_id).exists():
@@ -206,7 +207,7 @@ class RoomBlockList(APIView):
 
 
 class RoomUserBlock(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def post(self, request, room_id, user_id):
         if not room_exist(room_id):
@@ -245,7 +246,7 @@ class RoomUserBlock(APIView):
 
 
 class RoomUserUnBlock(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def delete(self, request, room_id, user_id):
         if not room_exist(room_id):
@@ -272,7 +273,7 @@ class RoomUserUnBlock(APIView):
 
 
 class RoomUserRemove(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def delete(self, request, room_id, user_id):
         if not room_exist(room_id):
@@ -295,7 +296,7 @@ class RoomUserRemove(APIView):
 
 
 class UserRoom(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request):
         rooms = Room.objects.filter(members__member=request.user)
@@ -304,7 +305,7 @@ class UserRoom(APIView):
 
 
 class SetAccessLevel(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def put(self, request, room_id):
         if not room_exist(room_id):
@@ -337,7 +338,7 @@ class SetAccessLevel(APIView):
 
 
 class TransferAdmin(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def put(self, request, room_id, user_id):
         if not room_exist(room_id):
@@ -360,7 +361,7 @@ class TransferAdmin(APIView):
 
 
 class InviteUser(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def post(self, request, room_id, username):
         if not CustomUser.objects.filter(username=username).exists():
@@ -392,7 +393,7 @@ class InviteUser(APIView):
 
 
 class AcceptInviting(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def post(self, request, invite_id):
         if not RoomInviting.objects.filter(id=invite_id, invited=request.user).exists():
@@ -414,7 +415,7 @@ class AcceptInviting(APIView):
 
 
 class RejectInviting(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def delete(self, request, invite_id):
         if not RoomInviting.objects.filter(id=invite_id, invited=request.user).exists():
@@ -426,7 +427,7 @@ class RejectInviting(APIView):
 
 
 class InvitationList(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request):
         invite_list = RoomInviting.objects.filter(invited=request.user)
@@ -435,7 +436,7 @@ class InvitationList(APIView):
 
 
 class RoomInvitationList(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsVerify]
 
     def get(self, request, room_id):
         if not room_exist(room_id):
